+++ mcramp/geom/banana.py
@@ -9,10 +9,10 @@ import re
 
 class GBanana(GPrim):
     def __init__(self, radius=0, height=0, mintheta=0, maxtheta=0, idx=0, ctx=None):
-        self.radius     = np.float32(radius)
-        self.height     = np.float32(height)
-        self.mintheta   = np.float32(mintheta)
-        self.maxtheta   = np.float32(maxtheta)
+        self.radius     = np.float64(radius)
+        self.height     = np.float64(height)
+        self.mintheta   = np.float64(mintheta)
+        self.maxtheta   = np.float64(maxtheta)
         self.idx        = idx
 
         with open(os.path.join(os.path.dirname(os.path.abspath(__file__)), 'banana.cl'), mode='r') as f:
